[Contract][Pets page] Add api contract
